pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "realtime-gdp-nowcast"
version = "0.1.0"
description = "Release-structured, revision-aware real-time U.S. GDP nowcasting research pipeline."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
  { name = "Codex" }
]
dependencies = [
  "beautifulsoup4>=4.12.3",
  "matplotlib>=3.10.1",
  "numpy>=2.2.5",
  "openpyxl>=3.1.5",
  "pandas>=2.2.3",
  "pyarrow>=19.0.1",
  "pyyaml>=6.0.2",
  "requests>=2.32.3",
  "scipy>=1.15.2",
  "seaborn>=0.13.2",
  "statsmodels>=0.14.4",
  "tabulate>=0.9.0",
  "tqdm>=4.67.1",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.5",
  "ruff>=0.11.8",
]

[project.scripts]
nowcast = "realtime_gdp_nowcast.cli:main"

[tool.hatch.build.targets.wheel]
only-include = ["src/realtime_gdp_nowcast"]

[tool.hatch.build.targets.wheel.sources]
"src" = ""

[tool.pytest.ini_options]
pythonpath = ["src"]
addopts = "-q"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
